Feature/adjust time

+++ src/core/tracking.rs
@@ -87,7 +87,71 @@ impl<'a> Tracking<'a> {
     /// # Errors
     ///
     /// Returns an error if `SQLite` fails to execute the queries.
-    pub fn set(&self, project_id: i32, date: Date, total_seconds: i64) -> rusqlite::Result<()> {
+    pub fn set(
+        &self,
+        project_id: i32,
+        date: Date,
+        total_seconds: i64,
+    ) -> Result<(), TrackingError> {
+        self.overwrite_day(project_id, date, Some(total_seconds))?;
+
+        Ok(())
+    }
+
+    /// Reset time tracking on the given project on the given date.
+    ///
+    /// It does so by removing any events and manual sessions on the given project on the given date.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if deleting or creating events in the database fails.
+    pub fn reset(&self, project_id: i32, date: Date) -> Result<(), TrackingError> {
+        self.overwrite_day(project_id, date, None)?;
+
+        Ok(())
+    }
+
+    /// Adjusts the recorded total time for a specific project's sessions to the nearest rounded
+    /// time (15-minute intervals)
+    ///
+    /// # Errors
+    ///
+    /// This method will return a `rusqlite::Error` in cases such as:
+    /// * Failure to retrieve sessions from the database.
+    /// * Failure to delete events from the `EventRepository`.
+    /// * Failure to upsert the adjusted session in the `ManualSessionRepository`.
+    pub fn adjust_by_fifteen_minutes(
+        &self,
+        project_id: i32,
+        date: Date,
+        operation: TimeAdjustmentOperation,
+    ) -> Result<(), TrackingError> {
+        const FIFTEEN_MINUTES_SECONDS: i64 = 15 * 60;
+
+        let sessions = self.list_all_sessions(date, Some(project_id))?;
+        let total_seconds = sessions.first().map_or(0, |s| s.total_seconds);
+
+        let rounded_seconds = match operation {
+            TimeAdjustmentOperation::Increment => {
+                ((total_seconds / FIFTEEN_MINUTES_SECONDS) + 1) * FIFTEEN_MINUTES_SECONDS
+            }
+            TimeAdjustmentOperation::Decrement => {
+                ((total_seconds - 1) / FIFTEEN_MINUTES_SECONDS) * FIFTEEN_MINUTES_SECONDS
+            }
+        }
+        .max(0);
+
+        self.overwrite_day(project_id, date, Some(rounded_seconds))?;
+
+        Ok(())
+    }
+
+    fn overwrite_day(
+        &self,
+        project_id: i32,
+        date: Date,
+        total_seconds: Option<i64>,
+    ) -> Result<(), TrackingError> {
         let event_repository = EventRepository::new(self.connection);
         let manual_session_repository = ManualSessionRepository::new(self.connection);
 
@@ -95,24 +159,11 @@ impl<'a> Tracking<'a> {
 
         // TODO: If a start event exists on the project on a previous date we need to handle it
 
-        manual_session_repository.upsert(project_id, date, total_seconds)?;
-
-        Ok(())
-    }
-
-    /// Reset time tracking on the given project on the given date.
-    ///
-    /// It does so by remove any events and manual sessions on the given project on the given date.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if deleting or creating events in the database fails.
-    pub fn reset(&self, project_id: i32, date: Date) -> rusqlite::Result<()> {
-        let event_repository = EventRepository::new(self.connection);
-        event_repository.delete_all_in(project_id, date)?;
-
-        let manual_session_repository = ManualSessionRepository::new(self.connection);
-        manual_session_repository.delete(project_id, date)?;
+        if let Some(seconds) = total_seconds {
+            manual_session_repository.upsert(project_id, date, seconds)?;
+        } else {
+            manual_session_repository.delete(project_id, date)?;
+        }
 
         Ok(())
     }
@@ -128,7 +179,11 @@ impl<'a> Tracking<'a> {
     /// # Errors
     ///
     /// Returns a database error if any query or project lookup fails.
-    pub fn list_all_sessions(&self, date: Date) -> rusqlite::Result<Vec<Session>> {
+    pub fn list_all_sessions(
+        &self,
+        date: Date,
+        project_id: Option<i32>,
+    ) -> rusqlite::Result<Vec<Session>> {
         struct PrimitiveSession {
             project_id: i32,
             total_seconds: i64,
@@ -141,7 +196,7 @@ impl<'a> Tracking<'a> {
 
         let mut project_to_primitive_sessions = HashMap::new();
 
-        manual_sessions_repository.for_each(date, |manual_session| {
+        manual_sessions_repository.for_each(date, project_id, |manual_session| {
             project_to_primitive_sessions.insert(
                 manual_session.project_id,
                 PrimitiveSession {
@@ -154,7 +209,7 @@ impl<'a> Tracking<'a> {
 
         let mut project_to_events: HashMap<i32, Vec<Event>> = HashMap::new();
 
-        event_repository.for_each(date, |event| {
+        event_repository.for_each(date, project_id, |event| {
             project_to_events
                 .entry(event.project_id)
                 .or_default()
@@ -223,6 +278,12 @@ impl<'a> Tracking<'a> {
 
         Ok(sessions)
     }
+}
+
+#[derive(Copy, Clone)]
+pub enum TimeAdjustmentOperation {
+    Increment,
+    Decrement,
 }
 
 #[derive(Debug, Error)]
@@ -420,9 +481,9 @@ mod tests {
             event_repository.insert(2, Stop, timestamp)?;
 
             // -------------------------
-            // Assert event based sessions duration
+            // Assert event-based sessions duration
             // -------------------------
-            let sessions = tracking.list_all_sessions(date)?;
+            let sessions = tracking.list_all_sessions(date, None)?;
             assert_eq!(sessions.len(), 2);
 
             let session = sessions.first().unwrap();
@@ -439,9 +500,9 @@ mod tests {
             tracking.set(1, date, 200)?;
 
             // -------------------------
-            // Verify that project 1 is 200 seconds, and project 2 is still event based
+            // Verify that project 1 is 200 seconds, and project 2 is still event-based
             // -------------------------
-            let sessions = tracking.list_all_sessions(date)?;
+            let sessions = tracking.list_all_sessions(date, None)?;
             assert_eq!(sessions.len(), 2);
 
             let session = sessions.first().unwrap();
@@ -477,11 +538,12 @@ mod tests {
             timestamp += 10;
             event_repository.insert(1, Stop, timestamp)?;
             timestamp += 10;
-            // Other project's events should not be deleted
+            // Other projects' events should not be deleted
             event_repository.insert(2, Start, timestamp)?;
 
             let next_day =
                 Date::from_calendar_date(2026, time::Month::June, 3).expect("valid date");
+
             let next_day_timestamp = PrimitiveDateTime::new(next_day, time)
                 .assume_utc()
                 .unix_timestamp();
@@ -510,6 +572,127 @@ mod tests {
 
             let sessions_after = context.collect_sessions()?;
             assert!(sessions_after.is_empty());
+
+            Ok(())
+        }
+    }
+
+    mod adjust_by_fifteen_minutes {
+        use super::*;
+        use crate::core::tracking::TimeAdjustmentOperation::Decrement;
+        use crate::core::tracking::TimeAdjustmentOperation::Increment;
+
+        #[test]
+        fn rounds_to_nearest_quarter() -> Result<(), Box<dyn Error>> {
+            let context = initialize_context()?;
+            let tracking = Tracking::new(context.connection());
+            let date = Date::from_calendar_date(2024, time::Month::April, 11)?;
+
+            tracking.set(1, date, 400)?;
+            tracking.set(2, date, 400)?;
+
+            // Rounding up to nearest quarter (900 seconds)
+            tracking.adjust_by_fifteen_minutes(1, date, Increment)?;
+            // Rounding down to nearest quarter (0 seconds)
+            tracking.adjust_by_fifteen_minutes(2, date, Decrement)?;
+
+            let sessions = tracking.list_all_sessions(date, None)?;
+            let first = sessions.first().unwrap();
+            let second = sessions.get(1).unwrap();
+
+            assert_eq!(first.project.id, 1);
+            assert!(!first.is_started);
+            assert_eq!(first.total_seconds, 900);
+
+            assert_eq!(second.project.id, 2);
+            assert!(!second.is_started);
+            assert_eq!(second.total_seconds, 0);
+
+            // Rounding up to 1,800 seconds
+            tracking.adjust_by_fifteen_minutes(1, date, Increment)?;
+            // Rounding up to 2,700 seconds
+            tracking.adjust_by_fifteen_minutes(1, date, Increment)?;
+            // Rounding down to 1,800 seconds
+            tracking.adjust_by_fifteen_minutes(1, date, Decrement)?;
+
+            let sessions = tracking.list_all_sessions(date, None)?;
+            let first = sessions.first().unwrap();
+            let second = sessions.get(1).unwrap();
+
+            assert_eq!(first.project.id, 1);
+            assert!(!first.is_started);
+            assert_eq!(first.total_seconds, 1800);
+
+            assert_eq!(second.project.id, 2);
+            assert!(!second.is_started);
+            assert_eq!(second.total_seconds, 0);
+
+            Ok(())
+        }
+
+        #[test]
+        fn zero_is_lowest_value() -> Result<(), Box<dyn Error>> {
+            let context = initialize_context()?;
+            let tracking = Tracking::new(context.connection());
+            let date = Date::from_calendar_date(2024, time::Month::April, 11)?;
+
+            tracking.set(1, date, 400)?;
+            // Rounding down to nearest quarter (0 seconds)
+            tracking.adjust_by_fifteen_minutes(1, date, Decrement)?;
+            // Rounding down again stays at 0 seconds.
+            tracking.adjust_by_fifteen_minutes(1, date, Decrement)?;
+
+            let sessions = tracking.list_all_sessions(date, None)?;
+            let session = sessions.first().unwrap();
+
+            assert_eq!(session.project.id, 1);
+            assert!(!session.is_started);
+            assert_eq!(session.total_seconds, 0);
+
+            Ok(())
+        }
+
+        #[test]
+        fn test_ongoing_sessions() -> Result<(), Box<dyn Error>> {
+            let context = initialize_context()?;
+            let tracking = Tracking::new(context.connection());
+            let date = OffsetDateTime::now_utc().date();
+
+            // Manual session (project 1)
+            tracking.set(1, date, 400)?;
+
+            // Ongoing session (project 2)
+            tracking.toggle(2)?;
+
+            // Adjusting a session does not modify the ongoing session
+            tracking.adjust_by_fifteen_minutes(1, date, Increment)?;
+
+            let sessions = tracking.list_all_sessions(date, None)?;
+            let first = sessions.first().unwrap();
+            let second = sessions.get(1).unwrap();
+
+            assert_eq!(first.project.id, 1);
+            assert!(!first.is_started);
+            assert_eq!(first.total_seconds, 900);
+
+            assert_eq!(second.project.id, 2);
+            assert!(second.is_started);
+            assert_eq!(second.total_seconds, 0);
+
+            // Adjusting an ongoing session stops it.
+            tracking.adjust_by_fifteen_minutes(2, date, Increment)?;
+
+            let sessions = tracking.list_all_sessions(date, None)?;
+            let first = sessions.first().unwrap();
+            let second = sessions.get(1).unwrap();
+
+            assert_eq!(first.project.id, 1);
+            assert!(!first.is_started);
+            assert_eq!(first.total_seconds, 900);
+
+            assert_eq!(second.project.id, 2);
+            assert!(!second.is_started);
+            assert_eq!(second.total_seconds, 900);
 
             Ok(())
         }
@@ -568,8 +751,8 @@ mod tests {
             // -------------------------
             // Execute
             // -------------------------
-            let sessions = tracking.list_all_sessions(date)?;
-            let today_sessions = tracking.list_all_sessions(today)?;
+            let sessions = tracking.list_all_sessions(date, None)?;
+            let today_sessions = tracking.list_all_sessions(today, None)?;
 
             // -------------------------
             // Assertions
@@ -641,7 +824,7 @@ mod tests {
             // -------------------------
             // Execute
             // -------------------------
-            let today_sessions = tracking.list_all_sessions(today)?;
+            let today_sessions = tracking.list_all_sessions(today, None)?;
 
             // -------------------------
             // Assertions
